requirements.txt
pandas 
numpy 
scikit-learn 
imbalanced-learn 
rdkit-pypi 
matplotlib 
seaborn 
xgboost 
lightgbm 
catboost